add:优化了数据管理模块的提示信息

+++ celloud/src/main/webapp/resources/js/data/controller.js
@@ -15,30 +15,61 @@
     $scope.conditionList = function(){
       $scope.pageQuery(1,$.dataManager.options.pageSize);
     }
+    //校验选择文件数量： 0 < num <= max
+    $scope.checkNum = function(){
+      var checkedIds = $.dataManager.options.checkedIds;
+      if(checkedIds.length==0){
+        $rootScope.errorInfo = "请选择至少一条记录";
+        $("#tips-modal").modal("show");
+        return false;
+      }
+      var maxNum = $.dataManager.options.dataMax;
+      if(checkedIds.length>maxNum){
+        $rootScope.errorInfo = "最多允许同时操作" + maxNum + "条数据！";
+        $("#tips-modal").modal("show");
+        return false;
+      }
+      return true;
+    }
     //项目运行
     $scope.runWithProject = function() {
+      if(!$scope.checkNum()){
+        return ;
+      }
+      if($.dataManager.options.noValidIds!=0){
+        $rootScope.errorInfo = "运行所勾选的数据必须有产品标签并且不处于运行状态！";
+        $("#tips-modal").modal("show");
+        return;
+      }
+      var checkedIds = $.dataManager.options.checkedIds;
+      if(checkedIds.length!=$.dataManager.options.validIds){
+        $rootScope.errorInfo = "数据选择异常，请刷新页面后重新选择！";
+        $("#tips-modal").modal("show");
+        return;
+      }
       runService.run().success(function(response) {
-        if(response.length==0){
-          alert("run success");
+        if(response.success){
           $scope.message = "运行成功";
-          $scope.state = true;
-          $scope.conditionList();
+          $scope.pageQuery($.dataManager.options.page,$.dataManager.options.pageSize);
           $.dataManager.refreshDataList();
         }else{
-          alert("run error");
+          $scope.message = response.message;
         }
+        $scope.state = true;
       });
     };
     //数据删除
     $scope.deleteData = function(){
+      if(!$scope.checkNum()){
+        return ;
+      }
       runService.delete().success(function(response){
-        if(response.code == "104"){
-          alert("success");
-          $scope.dataList = runService.list();
+        if(response.success){
+          $scope.conditionList();
           $.dataManager.refreshDataList();
-        }else{
-          alert("失败");
         }
+        $scope.message = response.message;
+        $scope.state = true;
       });
     };
     //跳转数据编辑
@@ -52,14 +83,13 @@
     $scope.submitEditData = function(){
       $scope.dataFile.tagName = $("input[type='radio'][name='dataTagName']:checked").val();
       runService.submitEditData($scope.dataFile).success(function(response){
-        if(response==1){
-          alert("success");
-          $scope.dataList = runService.list();
+        if(response.success){
+          $scope.pageQuery($.dataManager.options.page,$.dataManager.options.pageSize);
           $.dataManager.refreshDataList();
           $("#data-detail-modal").modal("hide");
-        }else{
-          alert("失败");
         }
+        $scope.message = response.message;
+        $scope.state = true;
       });
     }
   });
